Authentication

+++ backend/docker-start.sh
@@ -1,8 +1,14 @@
 #!/usr/bin/env sh
 set -e
 
-composer install --no-interaction --no-progress
-npm install
+if [ ! -f vendor/autoload.php ]; then
+	composer install --no-interaction --no-progress
+fi
+
+if [ ! -d node_modules ]; then
+	npm install
+fi
 php artisan migrate --force;
+php artisan db:seed --force;
 
 php artisan serve --host=0.0.0.0 --port=8000
